Separate current and obsolete formatting code
